x-pack/filebeat/input/http_endpoint: initial clean-up
