+++ x-pack/filebeat/input/http_endpoint/validate.go
@@ -10,15 +10,12 @@ import (
 	"crypto/sha1"
 	"crypto/sha256"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"hash"
 	"io"
 	"net/http"
 	"strings"
-
-	"github.com/elastic/elastic-agent-libs/logp"
 )
 
 var (
@@ -41,7 +38,7 @@ type apiValidator struct {
 	hmacPrefix         string
 }
 
-func (v *apiValidator) ValidateHeader(r *http.Request) (int, error) {
+func (v *apiValidator) validateRequest(r *http.Request) (status int, err error) {
 	if v.basicAuth {
 		username, password, _ := r.BasicAuth()
 		if v.username != username || v.password != password {
@@ -105,39 +102,5 @@ func (v *apiValidator) ValidateHeader(r *http.Request) (int, error) {
 		}
 	}
 
-	return 0, nil
-}
-
-type apiValidationHandler struct {
-	next      http.Handler
-	validator *apiValidator
-	log       *logp.Logger
-}
-
-func newAPIValidationHandler(next http.Handler, v *apiValidator, log *logp.Logger) http.Handler {
-	return &apiValidationHandler{
-		next:      next,
-		validator: v,
-		log:       log,
-	}
-}
-
-func (v *apiValidationHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if status, err := v.validator.ValidateHeader(r); status != 0 && err != nil {
-		sendAPIErrorResponse(w, r, v.log, status, err)
-		return
-	}
-
-	v.next.ServeHTTP(w, r)
-}
-
-func sendAPIErrorResponse(w http.ResponseWriter, r *http.Request, log *logp.Logger, status int, apiError error) {
-	w.Header().Add("Content-Type", "application/json")
-	w.WriteHeader(status)
-
-	enc := json.NewEncoder(w)
-	enc.SetEscapeHTML(false)
-	if err := enc.Encode(map[string]interface{}{"message": apiError.Error()}); err != nil {
-		log.Debugw("Failed to write HTTP response.", "error", err, "client.address", r.RemoteAddr)
-	}
+	return http.StatusAccepted, nil
 }
